add EvalContext.WithPath
As the Graph is walked, the current way to set the context path was to
have the walker return a context from EnterPath. This required that ever
node know it's absolute path, which can no longer be the case during
plan when modules have not been expanded.

This introduces a new method called WithPath, which returns a copy of
the context with the internal path updated to reflect the method
argument. This will let us set the path and properly evaluate
expressions within a DynamicExpand node, which was not a
GraphNodeInstance.